Move content and media controllers to their own namespaces
