+++ src/Umbraco.Cms.Api.Delivery/Controllers/Media/MediaApiControllerBase.cs
@@ -9,7 +9,7 @@ using Umbraco.Cms.Core.Services.OperationStatus;
 using Umbraco.Cms.Infrastructure.DeliveryApi;
 using Umbraco.Extensions;
 
-namespace Umbraco.Cms.Api.Delivery.Controllers;
+namespace Umbraco.Cms.Api.Delivery.Controllers.Media;
 
 [DeliveryApiMediaAccess]
 [VersionedDeliveryApiRoute("media")]
